new valgrind test for get_image_date

+++ util/examples/test_vast.sh
@@ -15100,6 +15100,23 @@ if [ $? -eq 0 ];then
      FAILED_TEST_CODES="$FAILED_TEST_CODES SPECIAL_VALGRIND029"
     fi
     #
+    valgrind -v --tool=memcheck --leak-check=full  --show-reachable=yes --track-origins=yes \
+    util/get_image_date '21/09/99' &> valgrind_test.out
+    if [ $? -ne 0 ];then
+     TEST_PASSED=0
+     FAILED_TEST_CODES="$FAILED_TEST_CODES SPECIAL_VALGRIND030"
+    fi
+    grep 'ERROR SUMMARY:' valgrind_test.out | awk -F ':' '{print $2}' | awk '{print $1}' | while read ERRORS ;do
+     if [ $ERRORS -ne 0 ];then
+      echo "ERROR"
+      break
+     fi
+    done | grep --quiet 'ERROR'
+    if [ $? -eq 0 ];then
+     TEST_PASSED=0
+     FAILED_TEST_CODES="$FAILED_TEST_CODES SPECIAL_VALGRIND031"
+    fi
+    #
     if [ ! -d ../vast_test_bright_stars_failed_match ];then
      cd ..
      wget -c "http://scan.sai.msu.ru/~kirx/pub/vast_test_bright_stars_failed_match.tar.bz2" && tar -xvjf vast_test_bright_stars_failed_match.tar.bz2 && rm -f vast_test_bright_stars_failed_match.tar.bz2
@@ -15112,7 +15129,7 @@ if [ $? -eq 0 ];then
      OMP_NUM_THREADS=1 valgrind -v --tool=memcheck --leak-check=full  --show-reachable=yes --track-origins=yes   ./vast -u -t2 -f ../vast_test_bright_stars_failed_match/* &> valgrind_test.out
      if [ $? -ne 0 ];then
       TEST_PASSED=0
-      FAILED_TEST_CODES="$FAILED_TEST_CODES SPECIAL_VALGRIND030"
+      FAILED_TEST_CODES="$FAILED_TEST_CODES SPECIAL_VALGRIND032"
      fi
      grep 'ERROR SUMMARY:' valgrind_test.out | awk -F ':' '{print $2}' | awk '{print $1}' | while read ERRORS ;do
       if [ $ERRORS -ne 0 ];then
@@ -15122,7 +15139,7 @@ if [ $? -eq 0 ];then
      done | grep --quiet 'ERROR'
      if [ $? -eq 0 ];then
       TEST_PASSED=0
-      FAILED_TEST_CODES="$FAILED_TEST_CODES SPECIAL_VALGRIND031"
+      FAILED_TEST_CODES="$FAILED_TEST_CODES SPECIAL_VALGRIND033"
      fi
     fi # if [ -d ../vast_test_bright_stars_failed_match ];then
     #
